[fix] remove Parallel() from T (BREAKING CHANGE) (#168)
The prior breaking change that added `.Parallel()` to `T` caused a
problem. (It broke something)
This reverts that part of the prior change and adds a `Parallel(t)`
function that calls an underlying `.Parallel()` if its supported.

Since we're already breaking things introduced in the last release,
`RunWithReWrap` has been replaced with `Run`. The new `Run` can handle
both `*testing.T` and `*testing.B` underlying types.