[chore] Replace AdjustSkipFrames with Helper (BREAKING CHANGE) (#170)
Since we're breaking things anyway, `AdjustSkipFrames` is gone in favor
or using `.Helper()`.
`AdjustSkipFrames` was just introduced so it can be removed w/o harming
anyone.

The `ReWrapper` interface has been changed to add `Unwrap() T`. This is
like unwrapping errors. This is another breaking change.

With this change, wrapped types, like what ReplaceLogger returns, do not
need to implement `.Run()` or `.Parallel()` and that allows them to work
with both an underlying `*testing.T` and a `*testing.B`.

`Run`, `Parallel` and `MustParallel` are all improved.